+++ packages/cli/src/process/spawn.ts
@@ -18,9 +18,6 @@ export default function (basePath: string, opts: Opts) {
 
     // Allows us to load an ESM module from a text string
     '--experimental-vm-modules',
-
-    // Allows us to do import('path/to/language-common') in the linker
-    '--experimental-specifier-resolution=node',
   ];
 
   const dirname = path.dirname(url.fileURLToPath(import.meta.url));