+++ packages/compiler/test/util/is-relative-specifier.test.ts
@@ -10,6 +10,7 @@ import { isRelativeSpecifier } from '../../src/util';
   '~/module',
   '.',
   '..',
+  'C:\\repo\\module', // windows absolute path
 ].forEach((path) => {
   test(`is a relative specifier: ${path}`, (t) => {
     t.truthy(isRelativeSpecifier(path));
@@ -23,6 +24,7 @@ import { isRelativeSpecifier } from '../../src/util';
   'namespace@module',
   'module@^3.2.1',
   '@namespace@module@=3.2.1',
+  'ab:\\repo\\module', // looks like a windows path, but is not!
 ].forEach((path) => {
   test(`is an absolute specifier: ${path}`, (t) => {
     t.falsy(isRelativeSpecifier(path));