+++ packages/runtime-manager/src/server/jobs/slow-random.js
@@ -1,5 +1,10 @@
 // This job takes a random number of seconds and returns a random number
-import { fn } from '@openfn/language-common';
+// import { fn } from '@openfn/language-common';
+
+// Note: since the linker no longer uses node resolution for import, common is unavailable
+// to the run-time manager. We basically need to add repo support.
+// For now, this simulates the behaviour well enough
+const fn = (f) => f;
 
 fn(
   (state) =>